Sections 5 & 7-VertexAI-Development/Bikeshare-Model/Cloud-Run-Flask-Online-Prediction/requirements.txt
Flask==2.1.0
gunicorn==20.1.0
google-cloud-aiplatform==1.26.0
pytest